Multiple fn version of Partition()
Partition() currently only accepts one predicate function and splits an
array in to two partitions, putting elements which pass in the first,
and those which fail in the second.

Partition() should accept n predicate functions, create n+1 partitions,
and place each element in the partition of the first predicate it
matches, with unmatched ones in the last.

In the case of single predicate function the behavior should be
unchanged.